+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 - Added getCACertificate and setCACertificate methods to HttpClientRequest.
+- Added getClientCertificate and setClientCertificate methods to HttpClientRequest.
 
 ## 1.0.0 - 2019-01-24
 - First stable revision.